third_party/blink/web_tests/fast/xmlhttprequest/xmlhttprequest-send-sharedarraybuffer.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

if (window.SharedArrayBuffer) {
  test(() => {
    const xhr = new XMLHttpRequest();
    xhr.open('POST', '/foo.html');

    assert_throws_js(TypeError, () => {
      xhr.send(new Uint8Array(new SharedArrayBuffer(32)));
    }, 'send() of SharedArrayBuffer view should throw');
  }, 'sending SharedArrayBuffer');
}
done();

</script>
